requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# Google Cloud dependencies
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
google-cloud-logging>=3.5.0
google-cloud-pubsub>=2.18.0
google-cloud-monitoring>=2.15.0

# Data processing
pyarrow>=12.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Web server dependencies
Flask>=2.3.0
gunicorn>=21.2.0

# Utilities
click>=8.1.0
requests>=2.31.0 